+++ CHANGELOG.md
@@ -910,6 +910,9 @@
   - AWSSNS
   - AWSPolly
 
+- **AWSS3**
+	- Added option to use Content-MD5 for multipart uploads with AWSS3TransferUtility (`useContentMD5` in `AWSS3TransferUtilityMultiPartUploadExpression`)
+
 ## 2.24.1
 
 ### Bug fixes